fix: data 파일 및 함수 폴더 위치 변경
